Paged search for matches we want to compute prediction quality for.

+++ footystats-tools-backend/src/main/java/de/ludwig/footystats/tools/backend/services/prediction/quality/Precast.java
@@ -12,5 +12,5 @@ public class Precast {
 
 	@Getter
 	@Setter
-	private Integer predictionsToAssess;
+	private Long predictionsToAssess;
 }
